// src/app/shared/model/duplicate-cust.model.ts
export class DuplicateCustObj {
  CustName: string;
  MrCustTypeCode: string;
  MrCustModelCode: string;
  MrIdTypeCode: string;
  IdNo: string;
  TaxIdNo: string;
  BirthDt: Date;
  MotherMaidenName: string;
  MobilePhnNo1: string;
  Status: string;
  Descr: string;
  IdExpiredDt: Date;
  RowVersion: string;
}